requirements.txt
torch
paddlepaddle
tensorflow
tf2onnx
onnx2tf
onnx2pytorch
x2paddle